backend/pab-submit/requirements.txt
psycopg2-binary==2.9.9
python-docx==1.1.0
boto3==1.34.0
